feat: add suspense and split Project Description with KPIs and move to components
