+++ front/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 
-import Footer from '@/components/Footer';
-import Navbar from '@/components/Navbar';
+import Footer from '@/app/components/Footer';
+import Navbar from '@/app/components/Navbar';
 import { Toaster } from '@/components/ui/toaster';
 
 import Providers from './Providers';
