+++ front/app/page.tsx
@@ -1,7 +1,7 @@
-import HomePageHeader from '@/components/HomePageHeader';
+import HomePageHeader from '@/app/components/HomePageHeader';
+import ProjectDescription from '@/app/components/ProjectDescription';
+import CtaGroup from '@/app/components/cta/CtaGroup';
 import FranceMap from '@/components/Map/open-tiles';
-import ProjectDescription from '@/components/ProjectDescription';
-import CtaGroup from '@/components/cta/CtaGroup';
 
 export default async function Home() {
   return (
